Remove unrar installation for Linux
Unrar seems to only play nice when certain files are in very specific
locations. Unfortunately these locations are owned by root. A fix could
be put together using sudo, but it makes more sense to install from apt
if sudo has to be used anyway. For now the README notes the unrar requirement
for Linux platforms.